clase modificada y documentada
