--- src/main/java/ed/camilo/Calculadora.java
@@ -1,22 +1,0 @@
-package ed.camilo;
-
-public class Calculadora {
-	
-	public int suma(int a, int b) {
-		return a + b;
-	}
-	
-	public int resta(int a, int b) {
-		return a - b;
-	}
-	
-	public int multiplica(int a, int b){
-		return a*b;
-	}
-	
-	public int diveix(int a, int b) {
-		return a/b;
-	}
-
-}
-
